% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_10_nouns.R
\name{get_top_10_nouns}
\alias{get_top_10_nouns}
\title{Title}
\usage{
get_top_10_nouns(start_date, end_date, type_speech, party, member, size)
}
\arguments{
\item{start_date}{the start of date(yyyy-mm-dd),or skip}

\item{end_date}{the end of date(yyyy-mm-dd),or skip}

\item{type_speech}{"Nej" or skip}

\item{party}{"c", "l", "kd", "mp", "m", "s", "sd", "v", "-", "nyd" or skip}

\item{member}{"id of member" or skip}

\item{size}{(10,50,200,1000,2000,5000,10000)}
}
\value{
first_10
}
\description{
Title
}
