% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/riksdagenAPI-package.R
\docType{package}
\name{riksdagenAPI-package}
\alias{riksdagenAPI}
\alias{riksdagenAPI-package}
\title{riksdagenAPI: Get Top 10 Nouns from Swedish Political Speeches}
\description{
This package retrieves the top 10 nouns from political speeches based on specific criteria such as date range, type of speech, party, member ID, and size of the response.
}
\keyword{internal}
