% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_top_5_nouns.R
\name{get_top_10_nouns}
\alias{get_top_10_nouns}
\title{Title}
\usage{
get_top_10_nouns(start_date, end_date, type_speech, party, member, size)
}
\arguments{
\item{start_date}{}

\item{end_date}{}

\item{type_speech}{}

\item{party}{}

\item{member}{}

\item{size}{}
}
\value{
first_10
}
\description{
Title
}
